+++ forui/CHANGELOG.md
@@ -25,6 +25,7 @@ We've done a pass over the controllers in Forui to make them more consistent and
 
 ### Others
 * Add `FTypeaheadController`.
+* Add `FPopover.onTapOutsideHide`.
 
 * Rename `FSelect.divider` to `FSelect.contentDivider`.
 * Change `FMultiValueNotifier` to be non-abstract.